app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Deep Sarker <deepsarker01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
mongoengine = "^0.29.1"
pymupdf = "^1.24.10"
nltk = "^3.9.1"
pypdf2 = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
